+++ src/abstractdataframe/reshape.jl
@@ -198,11 +198,11 @@ end
 
 """
     unstack(df::AbstractDataFrame, rowkeys, colkey, value; renamecols::Function=identity,
-            allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
+            allowmissing::Bool=false, allowduplicates::Bool=false, fill=missing)
     unstack(df::AbstractDataFrame, colkey, value; renamecols::Function=identity,
-            allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
+            allowmissing::Bool=false, allowduplicates::Bool=false, fill=missing)
     unstack(df::AbstractDataFrame; renamecols::Function=identity,
-            allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
+            allowmissing::Bool=false, allowduplicates::Bool=false, fill=missing)
 
 Unstack data frame `df`, i.e. convert it from long to wide format.
 
@@ -229,10 +229,10 @@ Row and column keys will be ordered in the order of their first appearance.
 - `allowduplicates`: if `false` (the default) then an error an error will be thrown
   if combination of `rowkeys` and `colkey` contains duplicate entries; if `true`
   then  then the last encountered `value` will be retained.
-- `fillvalue`: missing row/column combinations are filled with this value; The default
-  is `missing`. If the `value` column is a `CategoricalVector` and `fillvalue`
+- `fill`: missing row/column combinations are filled with this value; The default
+  is `missing`. If the `value` column is a `CategoricalVector` and `fill`
   is not `missing` then in order to keep unstacked value columns also
-  `CategoricalVector` the `fillvalue` must be passed as `CategoricalValue`
+  `CategoricalVector` the `fill` must be passed as `CategoricalValue`
 
 # Examples
 
@@ -347,7 +347,7 @@ julia> df = DataFrame(id=["1", "1", "2"],
    2 │ 1       Var2          2
    3 │ 2       Var1          3
 
-julia> unstack(df, :variable, :value, fillvalue=0)
+julia> unstack(df, :variable, :value, fill=0)
 2×3 DataFrame
  Row │ id      Var1   Var2
      │ String  Int64  Int64
@@ -359,7 +359,7 @@ Note that there are some differences between the widened results above.
 """
 function unstack(df::AbstractDataFrame, rowkeys, colkey::ColumnIndex,
                  value::ColumnIndex; renamecols::Function=identity,
-                 allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
+                 allowmissing::Bool=false, allowduplicates::Bool=false, fill=missing)
     rowkey_ints = vcat(index(df)[rowkeys])
     @assert rowkey_ints isa AbstractVector{Int}
     length(rowkey_ints) == 0 && throw(ArgumentError("No key column found"))
@@ -367,23 +367,23 @@ function unstack(df::AbstractDataFrame, rowkeys, colkey::ColumnIndex,
     g_colkey = groupby(df, colkey)
     valuecol = df[!, value]
     return _unstack(df, rowkey_ints, index(df)[colkey], g_colkey,
-                    valuecol, g_rowkey, renamecols, allowmissing, allowduplicates, fillvalue)
+                    valuecol, g_rowkey, renamecols, allowmissing, allowduplicates, fill)
 end
 
 function unstack(df::AbstractDataFrame, colkey::ColumnIndex, value::ColumnIndex;
                  renamecols::Function=identity,
-                 allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
+                 allowmissing::Bool=false, allowduplicates::Bool=false, fill=missing)
     colkey_int = index(df)[colkey]
     value_int = index(df)[value]
     return unstack(df, Not(colkey_int, value_int), colkey_int, value_int,
             renamecols=renamecols, allowmissing=allowmissing,
-            allowduplicates=allowduplicates, fillvalue=fillvalue)
+            allowduplicates=allowduplicates, fill=fill)
 end
 
 unstack(df::AbstractDataFrame; renamecols::Function=identity,
-        allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing) =
+        allowmissing::Bool=false, allowduplicates::Bool=false, fill=missing) =
     unstack(df, :variable, :value, renamecols=renamecols, allowmissing=allowmissing,
-            allowduplicates=allowduplicates, fillvalue=fillvalue)
+            allowduplicates=allowduplicates, fill=fill)
 
 # we take into account the fact that idx, starts and ends are computed lazily
 # so we rather directly reference the gdf.groups
@@ -411,7 +411,7 @@ function _unstack(df::AbstractDataFrame, rowkeys::AbstractVector{Int},
                   colkey::Int, g_colkey::GroupedDataFrame,
                   valuecol::AbstractVector, g_rowkey::GroupedDataFrame,
                   renamecols::Function,
-                  allowmissing::Bool, allowduplicates::Bool, fillvalue)
+                  allowmissing::Bool, allowduplicates::Bool, fill)
     rowref = g_rowkey.groups
     row_group_row_idxs = find_group_row(g_rowkey)
     Nrow = length(g_rowkey)
@@ -426,9 +426,9 @@ function _unstack(df::AbstractDataFrame, rowkeys::AbstractVector{Int},
                             "Pass `allowmissing=true` to skip missings."))
     end
     unstacked_val = [fill!(similar(valuecol,
-                                   promote_type(eltype(valuecol), typeof(fillvalue)),
+                                   promote_type(eltype(valuecol), typeof(fill)),
                                    Nrow),
-                           fillvalue) for _ in 1:Ncol]
+                           fill) for _ in 1:Ncol]
 
     mask_filled = falses(Nrow, Ncol)