+++ NEWS.md
@@ -22,10 +22,10 @@
   (notably `PooledArray` and `CategoricalArray`) or when they contained only
   integers in a small range.
   ([#2812](https://github.com/JuliaData/DataFrames.jl/pull/2812))
-* the `unstack` function receives new keyword argument `fillvalue`
+* the `unstack` function receives new keyword argument `fill`
   (with `missing` default) that is used to fill combinations of not encountered
   rows and columns. This feature allows to distinguish between missings in
-  value column and just missing row/column combinations and to easily fill 
+  value column and just missing row/column combinations and to easily fill
   with zeros non existing combinations in case of counting.
   ([#2828](https://github.com/JuliaData/DataFrames.jl/pull/2828))